+++ app/src/main/java/com/ebsolutions/papertrail/financialdataproviderservice/config/Constants.java
@@ -4,4 +4,5 @@ public class Constants {
   // URLs
   public static final String HEALTH_CHECK_URL = "/actuator/health";
   public static final String INFO_CHECK_URL = "/actuator/info";
+  public static final String USERS_URL = "/users";
 }