+++ app/src/main/java/com/ebsolutions/papertrail/financialdataproviderservice/user/UserController.java
@@ -1,0 +1,33 @@
+package com.ebsolutions.papertrail.financialdataproviderservice.user;
+
+import com.ebsolutions.papertrail.financialdataproviderservice.common.exception.DataProcessingException;
+import com.ebsolutions.papertrail.financialdataproviderservice.model.User;
+import java.util.List;
+import lombok.AllArgsConstructor;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+
+@Validated
+@RestController
+@AllArgsConstructor
+@RequestMapping("users")
+public class UserController {
+  private final UserRepository userRepository;
+
+  @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<?> getAll() {
+    try {
+      List<User> users = userRepository.readAll();
+
+      return !users.isEmpty() ? ResponseEntity.ok(users) : ResponseEntity.noContent().build();
+    } catch (DataProcessingException dpe) {
+      return ResponseEntity.internalServerError().body(dpe.getMessage());
+    }
+  }
+
+}